[MS] Moved importManager from src/common to src/services
